+++ apps/web/package.json
@@ -30,7 +30,6 @@
         "is-url": "^1.2.4",
         "isomorphic-unfetch": "^4.0.2",
         "lodash": "^4.17.21",
-        "masto": "^5.11.3",
         "misskey-js": "^0.0.16",
         "next": "^13.4.9",
         "notistack": "^3.0.1",
@@ -51,8 +50,7 @@
         "ui": "*",
         "services": "*",
         "utility-types": "^3.10.0",
-        "yup": "^1.2.0",
-        "zod": "^3.21.4"
+        "yup": "^1.2.0"
     },
     "devDependencies": {
         "@graphql-codegen/cli": "^4.0.1",
