+++ apps/web/src/components/Media/index.ts
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { PostAttachment } from "@services/types";
+import { PostAttachment } from "services";
 
 export interface MediaContextValues {
     openMediaViewer(attachments: PostAttachment[], index?: number): void;
