+++ packages/services/package.json
@@ -13,6 +13,7 @@
     },
     "devDependencies": {
         "@types/jest": "^29.5.3",
+        "@types/lodash": "^4.14.196",
         "eslint": "^8.44.0",
         "eslint-config-custom": "*",
         "jest": "^29.6.1",
@@ -21,5 +22,11 @@
         "tsconfig": "*",
         "tsup": "^7.1.0",
         "typescript": "^4.5.2"
+    },
+    "dependencies": {
+        "dayjs": "^1.11.9",
+        "lodash": "^4.17.21",
+        "masto": "^5.11.3",
+        "zod": "^3.21.4"
     }
 }
