+++ packages/services/src/mastodon/auth.types.ts
