+++ packages/services/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./events";
+export * from "./notifications";
